Exported Optimization Methods

Downhill.FixedRateDescent — Type

FixedRateDescent

Descent method which minimizes the objective function in the direction of antigradient at each step.

Downhill.MomentumDescent — Type

MomentumDescent

Descent method which minimizes the objective function in the direction of antigradient at each step.

Downhill.NesterovMomentum — Type

NesterovMomentum

Descent method which minimizes the objective function in the direction of antigradient at each step.

Downhill.SteepestDescent — Type

SteepestDescent

Descent method which minimizes the objective function in the direction of antigradient at each step.

Downhill.HyperGradDescent — Type

HyperGradDescent

Descent method which minimizes the objective function in the direction of antigradient at each step.

Downhill.CGDescent — Type

CGDescent

Conjugate gradient method (Hager-Zhang version [W.Hager, H.Zhang // SIAM J. Optim (2006) Vol. 16, pp. 170-192])

Downhill.BFGS — Type

BFGS

Quasi-Newton descent method.

Downhill.CholBFGS — Type

CholBFGS

Quasi-Newton descent method.