Basic Functions

DescentMethods.optimize! — Function

optimize!(M::Wrapper, fdf, x0)

Find an optimizer for fdf, starting with the initial approximation x0. fdf(x, g) must return a tuple (f(x), ∇f(x)) and, if g is mutable, overwrite it with the gradient.

optimize!(
    M::CoreMethod, fdf, x0;
    gtol=1e-6,
    maxiter=100,
    maxcalls=nothing,
    reset=true,
    constrain_step=nothing,
    log_stream=nothing
)

Find an optimizer for fdf, starting with the initial approximation x0.

Arguments:

M::CoreMethod: the core method to use for optimization

fdf(x, g)::Function: function to optimize. It must return a tuple (f(x), ∇f(x)) and, if g is mutable, overwrite it with the gradient.

x0: initial approximation

Keywords:

gtol::Real: stop optimization when the gradient norm is less

maxiter::Integer: force stop optimization after this number of iterations (use nothing or a negative value to not constrain iteration number)

maxcalls::Integer: force stop optimization after this number of function calls (use nothing or a negative value to not constrain call number)

reset=true: a value to pass as a keyword argument to the optimizer init! method

constrain_step(x0, d): a function to constrain step from x0 in the direction d. It must return a real-numbered value α such that x0 + αd is the maximum allowed step

log_stream::Union{IO,AbstractString}: IO stream or a file name to log the optimization process

DescentMethods.solver — Function

DescentMethods.solver(
    M::CoreMethod;
    gtol = 1e-6,
    maxiter = 100,
    maxcalls = nothing,
    constrain_step=nothing,
    log_stream=nothing
)

Return the wrapper object for a chosen method to solve an optimization problem with given parameters. For the description of keywords, see optimize!

DescentMethods.solver(M::DataType, x; gtol = convert(eltype(x0), 1e-6), maxiter = 100, maxcalls = nothing, constrain_step)

Return the wrapper object for a chosen method to solve an optimization problem with given parameters compatible with the dimensions of x.

DescentMethods.reset! — Function

reset!(M::DescentMethod, args...; kwargs...)

Reset the solver parameters to the default (or to specific value - see the documentation for the specific types).

Each method has to implement a parameter-free reset!(M) method.