Exported Optimization Methods

DescentMethods.FixedRateDescent — Type

FixedRateDescent

Descent method which minimizes the objective function in the direction of antigradient at each step.

DescentMethods.MomentumDescent — Type

MomentumDescent

Descent method which minimizes the objective function in the direction of antigradient at each step.

DescentMethods.NesterovMomentum — Type

NesterovMomentum

Descent method which minimizes the objective function in the direction of antigradient at each step.

DescentMethods.SteepestDescent — Type

SteepestDescent

Descent method which minimizes the objective function in the direction of antigradient at each step.

DescentMethods.HyperGradDescent — Type

HyperGradDescent

Descent method which minimizes the objective function in the direction of antigradient at each step.

DescentMethods.CGDescent — Type

CGDescent

Conjugate gradient method (Hager-Zhang version [W.Hager, H.Zhang // SIAM J. Optim (2006) Vol. 16, pp. 170-192])

DescentMethods.BFGS — Type

BFGS

Quasi-Newton descent method.

DescentMethods.CholBFGS — Type

CholBFGS

Quasi-Newton descent method.