Basic Functions

DescentMethods.optimize! — Function

optimize!(M::Wrapper, fdf, x0)

Find an optimizer for fdf, starting with the initial approximation x0. fdf(x, g) must return a tuple (f(x), ∇f(x)) and, if g is mutable, overwrite it with the gradient.

optimize!(M::CoreMethod, fdf, x0; gtol=1e-6, maxiter=100, maxcalls=nothing, reset=true, constrain_step=nothing)

Find an optimizer for fdf, starting with the initial approximation x0. fdf(x, g) must return a tuple (f(x), ∇f(x)) and, if g is mutable, overwrite it with the gradient. A function constrain_step(x0, d) may be passed to limit the step sizes.

DescentMethods.solver — Function

DescentMethods.solver(M::CoreMethod; gtol = convert(eltype(x0), 1e-6), maxiter = 100, maxcalls = nothing, constrain_step)

Return the wrapper object for a chosen method to solve an optimization problem with given parameters.

DescentMethods.solver(M::DataType, x; gtol = convert(eltype(x0), 1e-6), maxiter = 100, maxcalls = nothing, constrain_step)

Return the wrapper object for a chosen method to solve an optimization problem with given parameters compatible with the dimensions of x.

DescentMethods.reset! — Function

reset!(M::DescentMethod, args...; kwargs...)

Reset the solver parameters to the default (or to specific value - see the documentation for the specific types).

Each method has to implement a parameter-free reset!(M) method.